yc_edition/requirements.txt
pyowm==3.3.0
pyTelegramBotAPI==4.10.0
feedparser==6.0.10
pycbrf==1.1.0
openai==0.27.4
pyqrcode
pypng
ydb